app/requirements.txt
anyio==3.7.1
APScheduler==3.9.1.post1
async-timeout==4.0.2
redis==4.6.0
backports.zoneinfo==0.2.1
cachetools==5.2.1
certifi==2023.5.7
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
python-telegram-bot==20.0a4
pytz==2023.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzlocal==5.0.1